configs

+++ views/delete.php
@@ -1,5 +1,5 @@
 <?php 
-if(isset($_POST['delete_id'])){
+if(isset($_POST['delete_id']) && isset($_POST['image']) ){
     $existedArticle = new ArticleController();
     $existedArticle->deleteArticle();
 }else{
